requirements.txt
django==3.2
djangorestframework
django-cors-headers
mysqlclient
cloudinary
python-decouple
django-admin-tailwind
django-admin-autocomplete-list-filter
djangorestframework-simplejwt
drf-yasg
gunicorn